@@ -1,9 +1,10 @@
 package peterfajdiga.sszj.elements;
 
 import android.content.Context;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.util.TypedValue;
+
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 // source: https://stackoverflow.com/a/30256880
 
