@@ -1,10 +1,11 @@
 package peterfajdiga.sszj.elements;
 
 import android.content.Context;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.util.AttributeSet;
 import android.view.View;
+
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 public class WeightedLinearLayoutManager extends LinearLayoutManager {
 
