@@ -1,10 +1,11 @@
 package peterfajdiga.sszj.elements.adapters;
 
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 public class WordsAdapter extends RecyclerView.Adapter<WordsAdapter.WordViewHolder> {
 
