@@ -2,13 +2,8 @@ package peterfajdiga.sszj.sections;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.TextView;
 
-import peterfajdiga.sszj.R;
+import androidx.fragment.app.Fragment;
 
 /**
  * A simple {@link Fragment} subclass.
