@@ -2,10 +2,11 @@ package peterfajdiga.sszj.sections;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import peterfajdiga.sszj.R;
 import peterfajdiga.sszj.elements.GridAutofitLayoutManager;
